refactor : DB에서 조회한 Entity를 Line로 만드는 책임을 CommonService -> LineService로 위임
- CommonService 삭제
